Force expunge all instances when removing a service (#6836)
Summary:
so far we rely on an `OfferMatcherReconciler` to expunge reserved persistent instances after the reservation was destroyed. However, in some cases (e.g. Agent was removed and the task is `TASK_UNKNOWN`) this will never happen, leading to Marathon leaking an instance. With this fix we force expunge instances after killing all tasks *if the app is being removed by the user*

JIRA: DCOS-49521


Co-Authored-By: zen-dog <adukhovniy@mesosphere.io>