Use Java 8 for building.
Summary:
The change from mesosphere/mesos-plugin#23 installs Java 11 on our
Jenkins nodes. If the node does not terminate a Marathon build will end
up using Java 11 instead of Java 11. According to ToolsInfra nodes
should used only once. In DCOS_OSS-4605 I reported that this is not the
case but it was dismissed. The ticket has been reopened. Until that is
fixed we have to use this work around.