Queue instance queries with updates.
Summary:
A deployment cancellation can lead to a instance leak. This happens when
and update by the cancelled deployment is queued or saved in Zookeeper
while the new deployment, eg a rollback, is querying the instances from
the instance tracker. The instance tracker did not see the effect of the
update yet. So far queries by-pass updates. This change queues the
`ListBySpec` queries with instance update operations. This way a query
is guranteed to see all pending updates. We might still run into issues
when updates follow a query. Also, this works only on a run spec level
and not when all specs are queried.

JIRA issues: DCOS-51375