Introduced Linux seccomp for services (#6892)
Summary:
seccomp added for marathon apps, requires cluster configuration to run.  The seccomp for a UCR container is defined in LinuxInfo as per the design doc.
The LinuxInfo can be None which is expected if there is no seccomp defined.
Seccomp can be None, which is odd at this point as there is no other reason to have LinuxInfo at this point (although it is planned)
The seccomp can have profilename as None, in which case unconfined can be anything.. however it is only meaningful if unconfined is `True` in this situation.   `False` gives the expectation that the container will be confided by a profile.
A profile name being provided, the unconfined needs to be `False`

JIRA issues:  DCOS-50607,  DCOS-50608

Co-authored-by: Aleksey Dukhovniy <alex.dukhovniy@googlemail.com>
Co-authored-by: Karsten Jeschkies <k@jeschkies.xyz>