Restart legacy resident apps.  (#6920)
Summary:
The reservation id defaults to `SimplifiedId`. This breaks the restart of older
resident apps and pods since their reservations have an old id. Unfortunately,
we construct the reservation id with each match instead of persisting it. This
will change with MARATHON-8517 which requires a migration.

This patch uses the same logic from 1.7 and earlier to infer the reservation id
via `instance.appTask.reservationId`.

JIRA issues: DCOS_OSS-5212