Suppress and revive offers based on instance state
Summary:
Changed the ReviveOffersActor to react to instance state changes, instead of being triggered from within the launch queue. Every time an instance is scheduled, a reviveOffers call will be issued. Once all scheduled instances transitioned to another state, offers will be suppressed.

This commit did not adjust launch queue internal OffersWanted messages, but ignores them. It is also to be discussed if and how we want to handle revives when Marathon needs to unreserve/destroy.

JIRA issues: MARATHON-8627