Remove instance id from reservation id.
Summary:
This decouples reservation and instance ids for good and
will enable us to reassign reservations if required.

The `OfferMatcherReconciler` used the instance id to determine
whether the reservation should be destroyed. In Marathon 1.8 this
is only the case when the instance was removed. We can test this
by checking if there is any instance that holds the reservation.

Depends on #6921

JIRA issues: MARATHON-8517