Do not try to kill a terminal instance. (#6941)
Summary:
The `KillServiceActor` would queue a terminal instance for kill. The queued item would have no task
attached and thus would not issue any kill. The queue item would also never be dequeued. Thus
if a task transitioned from `TASK_UNKNOWN` to `TASK_RUNNING` the `KillServiceActor` would
ignore it.

This removes the queuing of terminal instances and issues a warning in the logs. We also verify
that the `KillStreamWatcher` reports a terminal instances as, well, terminal.

JIRA issues: MARATHON-8422