Persistent volumes with profile now default to DiskType.Mount (#6934)
Summary:
When using DC/OS Storage Service (DSS), users can provision storage volumes as disk resources tagged with a profile name. Marathon allows users to create services that make use of disks tagged with these profiles. The type of disk resource in Mesos will be Mount in this case, as this is the only disk type DSS supports. Therefore, Marathon should default to disk type Mount instead of Root, when a profileName is given for a persistent volume configuration.

JIRA issues: MARATHON-8631