Fix the rest of the tests
For simplicity, we flatten the snapshot with the instance update events so that
our scan will emit an extra event for our snapshot; this causes the
suppress/revive logic to have two total events, the zero event (empty
snapshot), then the event for the received snapshot. Diffing these two events
will result in a suppress or a revive event, where before the first suppress /
revive for the snapshot was not being sent.

We extract out the config object for simple parameters, and delegate the
passing of these config values to the LaunchQueueModule