Set a default value for offersWanted subjects
There are two sources for signal indicating if offers are wanted. Metronome is
based on Marathon and only needs one of those sources. By assigning a default
value to both subjects, the combineLatest mechanism can function and not wait
forever for both sources to provide an initial value before yielding combined
elements.