Revive and suppress offers based on instance state. (#6936)
Summary:
Only one business logic should decide whether to suppress or revives offers.

This patch encapsulates this logic in a simple stream that keeps track of a
set of instances that are either scheduled or waiting for their reservations to
be destroyed. It also counts the number of expunged resident instances.

If an instance is scheduled or needs its reservation to be destroyed the
logic will emit three revives to account for a possible race condition with
the offer matcher that might decline just the offer we want to revive.

If the number of expunged resident instances increases the logic will emit
one revive to try to have the reservation destroyed by the
`OfferMatcherReconciler`. A future improvement would use the Mesos'
[operator API](http://mesos.apache.org/documentation/latest/reservation/#operator-http-endpoints) to unreserve a reservation explicitly. Once that is done the
reservation id can be decoupled from instance ids. See MARATHON-8517
for details.

JIRA issues: MARATHON-8627

Co-authored-by: Matthias Eichstedt <matthias.eichstedt@mesosphere.io>
Co-authored-by: Tim Harper <tharper@mesosphere.com>