Fix bug where app-normalization logic treated relative paths as absolute
In the case of group-role enforcement, there was a bug in which posting an app
with id 'dev/service' was treated differently than '/dev/service'. This was due
to erroneous handling of relative vs absolute paths.

In this PR, we split out AbsolutePathId from PathId so that code modules can
express, via the type system, that they require an absolute path.

Other improvements were made by improving the consistency of the usage of the
word "absolute" and "root" when working with pathIds, which should alleviate
some confusion.

JIRA Issues: MARATHON-8669