Fix SI `test_declined_offer_due_to_resource_role`
previously, the `acceptedResourceRoles" = ["very-random-role"]` was removed by normalization (see `sanitize_accepted_resource_roles` deprecated feature which is enabled by default) so the app could deploy successfully. In order for the test to work, we need to deploy it into the proper top-level group (`/very-random-role`) using `role` with the same value.