Fix a super sneaky test bug
There was a very small chance that the RateLimiter async-scheduling behavior
would be applied for the very first element, and this would cause the first
element to be dropped from the buffer.

The bug was due to the fact that the clock may not change from one call to the next.