Migration to cull out invalid resource roles (#7031)
Previously, Marathon would allow acceptedResourceRoles to include role values other than "*" and the configured Mesos role. In these cases, Marathon would not launch the instances at all.

With the migration for storage-version 19.100, all services are assigned the role as configured by `--mesos_role`; With this migration (version 19.300), we remove all unrelated roles from `acceptedResourceRoles`. If this results in the removal of all values from `acceptedResourceRoles`, then the `acceptedResourceRoles` is defaulted to `["*"]`.

JIRA Issues: MARATHON-8654