Respect --new_group_enforce_role when implicitly creating groups (#7085)
When creating apps such as "/dev/sleeper" in a non-existent group "/dev", Marathon will auto-create the necessary groups in order to facilitate the request. Previously, this behavior would ignore the enforceRole defaulting configuration, instead opting to always auto-create said groups with enforceRole: false (and consequently to not use, by default, the group-role for the role /dev/sleeper).

In this PR, we rectify the situation by enclosing the intent for new default groups in RootGroup. This state is threaded through from update to update.

Additionally, we update the defaulting logic for roles to use the configured newGroupRoleDefaultBehavior parameter.

JIRA issues: MARATHON-8706