Fix default role for top-level enforceRole: false groups
When --new_group_enforce_role=top is configured, and a group has
enforceRole: false set, Marathon was applying the group-role as the
default for new services in said group. This behavior was inconsistent
with the original design and the documentation.

In this commit we bring the behavior into alignment with the
documentation, and add a test case to assert it.