Switch to use tools/packager makefile for docker images (#7096)
This allows us to build and publish docker images for previously built
and uploaded Marathon artifacts, similar to that which can be done for
native packages

Further, we update the packaging makefile to use a version prefix folder for all output files. Additionally, the packaging tests are parameterized so they can be run for any version of Marathon.