Include reschuduled failed instances in pod status. (#7110)
Summary:
Marathon does not show scheduled pod instances in its API. This is due
to backwards compatibility when scheduled instances did not exist.
However, a resident pod might be re-scheduled. These should be shown
with their terminal or failed state. This patch will change the
filtering so that scheduled pods that do have a history, ie agent info,
are considered as re-scheduled and thus included.

JIRA issues: MARATHON-8710