Fix crash loop of LaunchQueueActor (#7170)
When LaunchQueueActor would crash, it'd tear down the delay updates
stream, which would result in a Status.Failure with an
AbruptStageTerminationException message due to the actor being torn
down. The next instance of LaunchQueueActor would receive this, escalate
the exception and cause another crash, result then in its
materialization of delay updates to share the same fate.

This would repeat ad infinitum.